Add files via upload
I have developed a static clone version of Netflix using HTML, CSS, and JavaScript as part of my practice. This project replicates the design and layout of the Netflix website but does not include dynamic functionality such as video streaming. It was an exercise in front-end development, focusing on design accuracy and responsive user interface